requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg[binary]==3.2.3
dj-database-url==2.2.0
python-dotenv==1.0.1
whitenoise==6.7.0
requests==2.32.3
gunicorn==21.2.0
django-cors-headers==4.4.0
flake8==7.1.1
pytest==8.3.3
pytest-django==4.9.0
